fixes to create command
The following changes have been made:

In the create function , where it is trying to change the config file
for neo4j.conf and change the ports used, the older string
“dbms.connector.http.address=0.0.0.0:7474” is not used anymore and the
alternative “dbms.connector.http.listen_address=:7474” exists. I
changed the function appropriately to check and change the config file
based on the string used.

Using default names for ports:
At the end of the create function, it was checking if dbName is given
by the user. If yes, it would write multiple files and if not it
wouldn’t write any of them. This was preventing the code to properly
handle multiple databases because one of these files that wouldn’t not
have been written is “shell-port” which is used in the creation of the
next database instance.

The current version generates a name for databases if not determined in
input in the format “untitled<port-number>” and creates all files
regardless of the input. I explained this in the usage descriptions for
create